fix: incorrect border radius approach & unusual naming conventions
