+++ src/app/components/Board/Board.module.scss
@@ -1,17 +1,20 @@
 @import '../../../styles/variables';
 
+.boardContainer {
+  margin: 20px auto;
+  padding: 2px;
+  border: 2px solid rgba(250, 250, 250, 0.3);
+  border-radius: 12px;
+}
+
 .board {
   display: grid;
   grid-template-columns: repeat($default-board-size, $square-size-desktop); // Set by inline style now
   grid-template-rows: repeat($default-board-size, $square-size-desktop); // Set by inline style now
-  border: 2px solid rgba(250, 250, 250, 0.3);
   border-radius: 12px;
-  margin: 20px auto;
   width: fit-content;
-  padding: 2px;
 
-  overflow-clip-margin: 8px;
-
+  overflow: clip;
 
   @media (max-width: 768px) {
     grid-template-columns: repeat($default-board-size, $square-size-mobile); // Set by inline style
